refactoring: improved clarity of code and functions atomicity
